+++ drivers/pid.h
@@ -4,8 +4,8 @@
  *
  * This Library is licensed under a GPLv3 License
  **********************************************************************************************/
-#define AUTOMATIC  1
-#define MANUAL  0
+#define AUTOMATIC	1
+#define MANUAL	0
 #define DIRECT  0
 #define REVERSE  1
 
@@ -17,9 +17,10 @@ typedef struct
   float Kp;
   float Ki;
   float Kd;
-  short direction;
-  short sampleTime;
+  int direction;
+  int sampleTime;
 } PIDparams;
+
 
 /*Constructor (...)*********************************************************
  *    The parameters specified here are those for for which we can't set up
@@ -27,21 +28,22 @@ typedef struct
  ***************************************************************************/
 PIDinit(PIDparams params)
 {
-  SetOutputLimits(0, 100);        //default output limit corresponds to
-  //the arduino pwm limits
+	SetOutputLimits(0, 100);				//default output limit corresponds to
+	//the arduino pwm limits
 
-  SampleTime = 100;              //default Controller Sample Time is 0.1 seconds
+	SampleTime = 100;							//default Controller Sample Time is 0.1 seconds
 
-  PIDsetControllerDirection(params);
-  PIDsetTunings(params);
+	PIDsetControllerDirection(params);
+	PIDsetTunings(params);
 
-  lastTime = millis()-SampleTime;
-  inAuto = false;
-  myOutput = Output;
-  myInput = Input;
-  mySetpoint = Setpoint;
+	lastTime = millis()-SampleTime;
+	inAuto = false;
+	myOutput = Output;
+	myInput = Input;
+	mySetpoint = Setpoint;
 
 }
+
 
 /* Compute() **********************************************************************
  *     This, as they say, is where the magic happens.  this function should be called
@@ -56,7 +58,7 @@ void PIDCompute()
    if(timeChange>=SampleTime)
    {
       /*Compute all the working error variables*/
-    float input = *myInput;
+	  float input = *myInput;
       float error = *mySetpoint - input;
       ITerm+= (ki * error);
       if(ITerm > outMax) ITerm= outMax;
@@ -66,15 +68,16 @@ void PIDCompute()
       /*Compute PID Output*/
       float output = kp * error + ITerm- kd * dInput;
 
-    if(output > outMax) output = outMax;
+	  if(output > outMax) output = outMax;
       else if(output < outMin) output = outMin;
-    *myOutput = output;
+	  *myOutput = output;
 
       /*Remember some variables for next time*/
       lastInput = input;
       lastTime = now;
    }
 }
+
 
 /* SetTunings(...)*************************************************************
  * This function allows the controller's dynamic performance to be adjusted.
@@ -103,7 +106,7 @@ void PIDSetTunings(float Kp, float Ki, float Kd)
 /* SetSampleTime(...) *********************************************************
  * sets the period, in Milliseconds, at which the calculation is performed
  ******************************************************************************/
-void PIDSetSampleTime(short NewSampleTime)
+void PIDSetSampleTime(int NewSampleTime)
 {
    if (NewSampleTime > 0)
    {
@@ -131,11 +134,11 @@ void PIDSetOutputLimits(float Min, float Max)
 
    if(inAuto)
    {
-     if(*myOutput > outMax) *myOutput = outMax;
-     else if(*myOutput < outMin) *myOutput = outMin;
+	   if(*myOutput > outMax) *myOutput = outMax;
+	   else if(*myOutput < outMin) *myOutput = outMin;
 
-     if(ITerm > outMax) ITerm= outMax;
-     else if(ITerm < outMin) ITerm= outMin;
+	   if(ITerm > outMax) ITerm= outMax;
+	   else if(ITerm < outMin) ITerm= outMin;
    }
 }
 
@@ -144,7 +147,7 @@ void PIDSetOutputLimits(float Min, float Max)
  * when the transition from manual to auto occurs, the controller is
  * automatically initialized
  ******************************************************************************/
-void PIDSetMode(short Mode)
+void PIDSetMode(int Mode)
 {
     bool newAuto = (Mode == AUTOMATIC);
     if(newAuto == !inAuto)
@@ -155,7 +158,7 @@ void PIDSetMode(short Mode)
 }
 
 /* Initialize()****************************************************************
- *  does all the things that need to happen to ensure a bumpless transfer
+ *	does all the things that need to happen to ensure a bumpless transfer
  *  from manual to automatic mode.
  ******************************************************************************/
 void PIDInitialize()
@@ -172,11 +175,11 @@ void PIDInitialize()
  * know which one, because otherwise we may increase the output when we should
  * be decreasing.  This is called from the constructor.
  ******************************************************************************/
-void PIDSetControllerDirection(short Direction)
+void PIDSetControllerDirection(int Direction)
 {
    if(inAuto && Direction !=controllerDirection)
    {
-    kp = (0 - kp);
+	  kp = (0 - kp);
       ki = (0 - ki);
       kd = (0 - kd);
    }
